Refactor crossmemory.c
1. Add LPA wrappers
2. Add ISGENQ wrappers
3. Add a source file for various cross-memory utility functions (cm memcpy)

Resolves zowe/zss#13
