+++ c/crossmemory.c
@@ -29,7 +29,9 @@
 
 #include "zowetypes.h"
 #include "alloc.h"
+#include "cmutils.h"
 #include "crossmemory.h"
+#include "lpa.h"
 #include "nametoken.h"
 #include "zos.h"
 #include "printables_for_dump.h"
@@ -41,10 +43,6 @@
 
 #define CROSS_MEMORY_SERVER_MIN_NAME_LENGTH   4
 #define CROSS_MEMORY_SERVER_MAX_NAME_LENGTH   16
-
-#ifndef CROSS_MEMORY_SERVER_KEY
-#define CROSS_MEMORY_SERVER_KEY 4
-#endif
 
 #ifndef CROSS_MEMORY_SERVER_RACF_PROFILE
 #define CROSS_MEMORY_SERVER_RACF_PROFILE CMS_PROD_ID".IS"
@@ -93,17 +91,6 @@ static void eightCharStringInit(EightCharString *string, char *cstring, char pad
 
 ZOWE_PRAGMA_PACK
 
-typedef struct QName_tag {
-  char value[8];
-} QName;
-
-typedef struct RName_tag {
-  unsigned char length;
-  char value[255];
-} RName;
-
-ZOWE_PRAGMA_PACK_RESET
-
 /* This is the QNAME used for all product enqueues. It must NEVER change. */
 static const QName PRODUCT_QNAME   = {CMS_PROD_ID"    "};
 /* These QNAME and RNAME must NEVER change and be the same across all
@@ -113,7 +100,7 @@ static const RName ZVTE_RNAME  = {8, "ISZVTE  "};
 
 #define CMS_MAX_ZVTE_CHAIN_LENGTH   128
 
-#ifdef METTLE
+#if defined(METTLE) && !defined(CMS_CLIENT)
 
 ZOWE_PRAGMA_PACK
 
@@ -306,7 +293,7 @@ void cmsInitializeLogging() {
 
 }
 
-#endif /* METTLE */
+#endif /* defined(METTLE) && !defined(CMS_CLIENT) */
 
 ZOWE_PRAGMA_PACK
 
@@ -581,211 +568,6 @@ static char *dumpWithEmptyMessageID(char *workBuffer, int workBufferSize, void *
   return NULL;
 }
 
-/* scope values can be found in SYS1.MACLIB(ISGYENQ) */
-#define ISGENQ_SCOPE_STEP     1
-#define ISGENQ_SCOPE_SYSTEM   2
-#define ISGENQ_SCOPE_SYSTEMS  3
-#define ISGENQ_SCOPE_SYSPLEX  4
-
-#ifdef METTLE
-__asm("GLBENQPL    ISGENQ MF=(L,GLBENQPL)" : "DS"(GLBENQPL));
-#endif
-
-static int isgenqGetExclusiveLockOrFail(const QName *qname, const RName *rname, uint8_t scope, ENQToken *token, int *reasonCode) {
-
-  QName localQName = *qname;
-  RName localRName = *rname;
-
-#ifdef METTLE
-  __asm(" ISGENQ MF=L" : "DS"(isgenqParmList));
-  isgenqParmList = GLBENQPL;
-#else
-  char isgenqParmList[512];
-#endif
-
-  int rc = 0, rsn = 0;
-  __asm(
-      ASM_PREFIX
-      "         PUSH USING                                                     \n"
-      "         DROP                                                           \n"
-      "ENQOBTX  LARL  2,ENQOBTX                                                \n"
-      "         USING ENQOBTX,2                                                \n"
-      "         ISGENQ   REQUEST=OBTAIN"
-      ",TEST=NO"
-      ",CONTENTIONACT=FAIL"
-      ",USERDATA=NO_USERDATA"
-      ",RESLIST=NO"
-      ",QNAME=(%2)"
-      ",RNAME=(%3)"
-      ",RNAMELEN=(%4)"
-      ",CONTROL=EXCLUSIVE"
-      ",RESERVEVOLUME=NO"
-      ",SCOPE=VALUE"
-      ",SCOPEVAL=(%5)"
-      ",RNL=YES"
-      ",ENQTOKEN=(%6)"
-      ",COND=YES"
-      ",RETCODE=%0"
-      ",RSNCODE=%1"
-      ",PLISTVER=IMPLIED_VERSION"
-      ",MF=(E,(%7),COMPLETE)"
-      "                                                                        \n"
-      "         DROP                                                           \n"
-      "         POP USING                                                      \n"
-      : "=m"(rc), "=m"(rsn)
-      : "r"(&localQName.value), "r"(&localRName.value), "r"(&localRName.length), "r"(&scope), "r"(token), "r"(&isgenqParmList)
-      : "r0", "r1", "r2", "r14", "r15"
-  );
-
-  if (reasonCode != NULL) {
-    *reasonCode = rsn;
-  }
-  return rc;
-}
-
-static int isgenqGetExclusiveLock(const QName *qname, const RName *rname, uint8_t scope, ENQToken *token, int *reasonCode) {
-
-  QName localQName = *qname;
-  RName localRName = *rname;
-
-#ifdef METTLE
-  __asm(" ISGENQ MF=L" : "DS"(isgenqParmList));
-  isgenqParmList = GLBENQPL;
-#else
-  char isgenqParmList[512];
-#endif
-
-  int rc = 0, rsn = 0;
-  __asm(
-      ASM_PREFIX
-      "         PUSH USING                                                     \n"
-      "         DROP                                                           \n"
-      "ENQOBTXW LARL  2,ENQOBTXW                                               \n"
-      "         USING ENQOBTXW,2                                               \n"
-      "         ISGENQ   REQUEST=OBTAIN"
-      ",TEST=NO"
-      ",CONTENTIONACT=WAIT"
-      ",USERDATA=NO_USERDATA"
-      ",RESLIST=NO"
-      ",QNAME=(%2)"
-      ",RNAME=(%3)"
-      ",RNAMELEN=(%4)"
-      ",CONTROL=EXCLUSIVE"
-      ",RESERVEVOLUME=NO"
-      ",SCOPE=VALUE"
-      ",SCOPEVAL=(%5)"
-      ",RNL=YES"
-      ",ENQTOKEN=(%6)"
-      ",COND=YES"
-      ",RETCODE=%0"
-      ",RSNCODE=%1"
-      ",PLISTVER=IMPLIED_VERSION"
-      ",MF=(E,(%7),COMPLETE)"
-      "                                                                        \n"
-      "         DROP                                                           \n"
-      "         POP USING                                                      \n"
-      : "=m"(rc), "=m"(rsn)
-      : "r"(&localQName.value), "r"(&localRName.value), "r"(&localRName.length), "r"(&scope), "r"(token), "r"(&isgenqParmList)
-      : "r0", "r1", "r2", "r14", "r15"
-  );
-
-  if (reasonCode != NULL) {
-    *reasonCode = rsn;
-  }
-  return rc;
-}
-
-static int isgenqTestLock(const QName *qname, const RName *rname, uint8_t scope, int *reasonCode) {
-
-  QName localQName = *qname;
-  RName localRName = *rname;
-
-#ifdef METTLE
-  __asm(" ISGENQ MF=L" : "DS"(isgenqParmList));
-  isgenqParmList = GLBENQPL;
-#else
-  char isgenqParmList[512];
-#endif
-
-  ENQToken localToken;
-
-  int rc = 0, rsn = 0;
-  __asm(
-      ASM_PREFIX
-      "         PUSH USING                                                     \n"
-      "         DROP                                                           \n"
-      "ENQOBTXT LARL  2,ENQOBTXT                                               \n"
-      "         USING ENQOBTXT,2                                               \n"
-      "         ISGENQ   REQUEST=OBTAIN"
-      ",TEST=YES"
-      ",RESLIST=NO"
-      ",QNAME=(%2)"
-      ",RNAME=(%3)"
-      ",RNAMELEN=(%4)"
-      ",CONTROL=EXCLUSIVE"
-      ",RESERVEVOLUME=NO"
-      ",SCOPE=VALUE"
-      ",SCOPEVAL=(%5)"
-      ",RNL=YES"
-      ",ENQTOKEN=(%6)"
-      ",COND=YES"
-      ",RETCODE=%0"
-      ",RSNCODE=%1"
-      ",PLISTVER=IMPLIED_VERSION"
-      ",MF=(E,(%7),COMPLETE)"
-      "                                                                        \n"
-      "         DROP                                                           \n"
-      "         POP USING                                                      \n"
-      : "=m"(rc), "=m"(rsn)
-      : "r"(&localQName.value), "r"(&localRName.value), "r"(&localRName.length), "r"(&scope), "r"(&localToken), "r"(&isgenqParmList)
-      : "r0", "r1", "r2", "r14", "r15"
-  );
-
-  if (reasonCode != NULL) {
-    *reasonCode = rsn;
-  }
-  return rc;
-}
-
-static int isgenqReleaseLock(ENQToken *token, int *reasonCode) {
-
-#ifdef METTLE
-  __asm(" ISGENQ MF=L" : "DS"(isgenqParmList));
-  isgenqParmList = GLBENQPL;
-#else
-  char isgenqParmList[512];
-#endif
-
-  int rc = 0, rsn = 0;
-  __asm(
-      ASM_PREFIX
-      "         PUSH USING                                                     \n"
-      "         DROP                                                           \n"
-      "ENQREL   LARL  2,ENQREL                                                 \n"
-      "         USING ENQREL,2                                                 \n"
-      "         ISGENQ   REQUEST=RELEASE"
-      ",RESLIST=NO"
-      ",ENQTOKEN=(%2)"
-      ",OWNINGTTOKEN=CURRENT_TASK"
-      ",COND=YES"
-      ",RETCODE=%0"
-      ",RSNCODE=%1"
-      ",PLISTVER=IMPLIED_VERSION"
-      ",MF=(E,(%3),COMPLETE)"
-      "                                                                        \n"
-      "         DROP                                                           \n"
-      "         POP USING                                                      \n"
-      : "=m"(rc), "=m"(rsn)
-      : "r"(token), "r"(&isgenqParmList)
-      : "r0", "r1", "r2", "r14", "r15"
-  );
-
-  if (reasonCode != NULL) {
-    *reasonCode = rsn;
-  }
-  return rc;
-}
-
 static int callPCRoutine(int pcNumber, int sequenceNumber, void *parmBlock) {
 
   int returnCode = 0;
@@ -805,7 +587,7 @@ static int callPCRoutine(int pcNumber, int sequenceNumber, void *parmBlock) {
   return returnCode;
 }
 
-#ifdef METTLE
+#if defined(METTLE) && !defined(CMS_CLIENT)
 
 static int axset() {
 
@@ -1064,183 +846,6 @@ static void wait(ECB * __ptr32 ecb) {
   );
 }
 
-static int getCallersKey() {
-
-  int key = 0;
-  __asm(
-      ASM_PREFIX
-      "         LA    7,1                                                      \n"
-      "         ESTA  6,7                                                      \n"
-      "         SRL   6,20                                                     \n"
-      "         N     6,=X'0000000F'                                           \n"
-      "         ST    6,%0                                                     \n"
-      : "=m"(key)
-      :
-      : "r6", "r7"
-  );
-
-  return key;
-}
-
-#define CROSS_MEMORY_ALET_PASN  0
-#define CROSS_MEMORY_ALET_SASN  1
-#define CROSS_MEMORY_ALET_HASN  2
-
-static void copyWithDestinationKey(void *dest,
-                                   unsigned int destKey,
-                                   unsigned int destALET,
-                                   const void *src,
-                                   size_t size) {
-
-  __asm(
-      ASM_PREFIX
-      /* input parameters */
-      "         LA    2,0(,%0)            destination address                  \n"
-      "         LA    4,0(,%1)            source address                       \n"
-      "         LA    1,0(,%2)            destination key                      \n"
-      "         SLL   1,4                 shift key to bits 24-27              \n"
-      "         LA    5,0                 use r5 for 0 value                   \n"
-      "         SAR   4,5                 source is primary (0)                \n"
-      "         SAR   2,%3                destination address space type       \n"
-      "         LA    5,0(,%4)            size                                 \n"
-      /* establish addressability */
-      "         PUSH  USING                                                    \n"
-      "         DROP                                                           \n"
-      "LCMCTBGN DS    0H                                                       \n"
-      "         LARL  10,LCMCTBGN                                              \n"
-      "         USING LCMCTBGN,10                                              \n"
-      /* enable AR mode for MVCDK */
-      "         SAC   512                 AR mode                              \n"
-      /* copy data in a loop */
-      "LCMTCPY1 DS    0H                                                       \n"
-#ifdef _LP64
-      "         LTGR  5,5                 size > 0?                            \n"
-#else
-      "         LTR   5,5                                                      \n"
-#endif
-      "         BNP   LCMTEXIT            no, leave                            \n"
-      "         LA    0,256                                                    \n"
-#ifdef _LP64
-      "         CGR   5,0                 size left > 256?                     \n"
-#else
-      "         CR    5,0                                                      \n"
-#endif
-      "         BH    LCMTCPY2            yes, go subtract 256                 \n"
-#ifdef _LP64
-      "         LGR   0,5                 no, load bytes left to r0            \n"
-#else
-      "         LR    0,5                                                      \n"
-#endif
-      "LCMTCPY2 DS    0H                                                       \n"
-#ifdef _LP64
-      "         SGR   5,0                 subtract 256                         \n"
-#else
-      "         SR    5,0                                                      \n"
-#endif
-      "         BCTR  0,0                 bytes to copy - 1                    \n"
-      "         MVCDK 0(2),0(4)           copy                                 \n"
-      "         LA    2,256(,2)           bump destination address             \n"
-      "         LA    4,256(,4)           bump source address                  \n"
-      "         B     LCMTCPY1            repeat                               \n"
-      /* go back into primary mode and leave */
-      "LCMTEXIT DS    0H                                                       \n"
-      "         SAC   0                   go back to primary mode              \n"
-      "         POP   USING                                                    \n"
-      :
-      : "r"(dest), "r"(src), "r"(destKey), "r"(destALET), "r"(size)
-      : "r0", "r1", "r2", "r4", "r5", "r10"
-  );
-
-}
-
-static void copyWithSourceKey(void *dest,
-                              const void *src,
-                              unsigned int srcKey,
-                              unsigned int srcALET,
-                              size_t size) {
-
-  __asm(
-      ASM_PREFIX
-      /* input parameters */
-      "         LA    2,0(,%0)            destination address                  \n"
-      "         LA    4,0(,%1)            source address                       \n"
-      "         LA    1,0(,%2)            source key                           \n"
-      "         SLL   1,4                 shift key to bits 24-27              \n"
-      "         LA    5,0                 use r5 for 0 value                   \n"
-      "         SAR   2,5                 destination is primary (0)           \n"
-      "         SAR   4,%3                source address ALET                  \n"
-      "         LA    5,0(,%4)            size                                 \n"
-      /* establish addressability */
-      "         PUSH  USING                                                    \n"
-      "         DROP                                                           \n"
-      "LCMCFBGN DS    0H                                                       \n"
-      "         LARL  10,LCMCFBGN                                              \n"
-      "         USING LCMCFBGN,10                                              \n"
-      /* enable AR mode for MVCDK */
-      "         SAC   512                 AR mode                              \n"
-      /* copy data in a loop */
-      "LCMFCPY1 DS    0H                                                       \n"
-#ifdef _LP64
-      "         LTGR  5,5                 size > 0?                            \n"
-#else
-      "         LTR   5,5                                                      \n"
-#endif
-      "         BNP   LCMFEXIT            no, leave                            \n"
-      "         LA    0,256                                                    \n"
-#ifdef _LP64
-      "         CGR   5,0                 size left > 256?                     \n"
-#else
-      "         CR    5,0                                                      \n"
-#endif
-      "         BH    LCMFCPY2            yes, go subtract 256                 \n"
-#ifdef _LP64
-      "         LGR   0,5                 no, load bytes left to r0            \n"
-#else
-      "         LR    0,5                                                      \n"
-#endif
-      "LCMFCPY2 DS    0H                                                       \n"
-#ifdef _LP64
-      "         SGR   5,0                 subtract 256                         \n"
-#else
-      "         SR    5,0                                                      \n"
-#endif
-      "         BCTR  0,0                 bytes to copy - 1                    \n"
-      "         MVCSK 0(2),0(4)           copy                                 \n"
-      "         LA    2,256(,2)           bump destination address             \n"
-      "         LA    4,256(,4)           bump source address                  \n"
-      "         B     LCMFCPY1            repeat                               \n"
-      /* go back into primary mode and leave */
-      "LCMFEXIT DS    0H                                                       \n"
-      "         SAC   0                   go back to primary mode              \n"
-      "         POP   USING                                                    \n"
-      :
-      : "r"(dest), "r"(src), "r"(srcKey), "r"(srcALET), "r"(size)
-      : "r0", "r1", "r2", "r4", "r5", "r10"
-  );
-
-}
-
-void cmCopyToSecondaryWithCallerKey(void *dest, const void *src, size_t size) {
-  copyWithDestinationKey(dest, getCallersKey(), CROSS_MEMORY_ALET_SASN, src, size);
-}
-
-void cmCopyFromSecondaryWithCallerKey(void *dest, const void *src, size_t size) {
-  copyWithSourceKey(dest, src, getCallersKey(), CROSS_MEMORY_ALET_SASN, size);
-}
-
-void cmCopyToPrimaryWithCallerKey(void *dest, const void *src, size_t size) {
-  copyWithDestinationKey(dest, getCallersKey(), CROSS_MEMORY_ALET_PASN, src, size);
-}
-
-void cmCopyFromPrimaryWithCallerKey(void *dest, const void *src, size_t size) {
-  copyWithSourceKey(dest, src, getCallersKey(), CROSS_MEMORY_ALET_PASN, size);
-}
-
-void cmCopyWithSourceKeyAndALET(void *dest, const void *src, unsigned int key,
-                                unsigned int alet, size_t size) {
-  copyWithSourceKey(dest, src, key, alet, size);
-}
-
 int cmsAddConfigParm(CrossMemoryServer *server,
                      const char *name, const void *value,
                      CrossMemoryServerParmType type) {
@@ -1321,7 +926,8 @@ static void racfEntityNameInit(RACFEnityName *string, char *cstring, char padCha
 
 static bool isSRBCaller() {
   PSA callerPSA;
-  copyWithSourceKey(&callerPSA, 0x00, 0, CROSS_MEMORY_ALET_HASN, sizeof(PSA));
+  cmCopyWithSourceKeyAndALET(&callerPSA, 0x00, 0, CROSS_MEMORY_ALET_HASN,
+                             sizeof(PSA));
   return callerPSA.psatold != NULL ? FALSE : TRUE;
 }
 
@@ -1401,109 +1007,6 @@ static int getTCBKey() {
   TCB *tcb = getTCB();
   int key = ((int)tcb->tcbpkf >> 4) & 0x0000000F;
   return key;
-}
-
-static void getCallerTCB(TCB *content, TCB **address) {
-
-  TCB * __ptr32 callerTCBAddr = NULL;
-  copyWithSourceKey(&callerTCBAddr, (void *)CURRENT_TCB, 0, CROSS_MEMORY_ALET_HASN, 4);
-  copyWithSourceKey(content, callerTCBAddr, 0, CROSS_MEMORY_ALET_HASN, sizeof(TCB));
-
-  if (address != NULL) {
-    *address = callerTCBAddr;
-  }
-
-}
-
-static void getCallerASCB(ASCB *content, ASCB **address) {
-
-  ASCB * __ptr32 callerASCBAddr = NULL;
-  copyWithSourceKey(&callerASCBAddr, (void *)CURRENT_ASCB, 0, CROSS_MEMORY_ALET_HASN, 4);
-  copyWithSourceKey(content, callerASCBAddr, 0, CROSS_MEMORY_ALET_HASN, sizeof(ASCB));
-
-  if (address != NULL) {
-    *address = callerASCBAddr;
-  }
-
-}
-
-static void getCallerASXB(ASXB *content, ASXB **address) {
-
-  ASCB callerASCB;
-  getCallerASCB(&callerASCB, NULL);
-
-  ASXB * __ptr32 callerASXBAddr = callerASCB.ascbasxb;
-  copyWithSourceKey(content, callerASXBAddr, 0, CROSS_MEMORY_ALET_HASN, sizeof(ASXB));
-
-  if (address != NULL) {
-    *address = callerASXBAddr;
-  }
-
-}
-
-static void getCallerACEE(ACEE *content, ACEE **address) {
-
-  TCB callerTCB;
-  getCallerTCB(&callerTCB, NULL);
-
-  ACEE * __ptr32 callerACEEAddr = callerTCB.tcbsenv;
-
-  if (callerACEEAddr != NULL) {
-
-    copyWithSourceKey(content, callerACEEAddr, 0, CROSS_MEMORY_ALET_HASN, sizeof(ACEE));
-
-  } else {
-
-    ASXB callerASXB;
-    getCallerASXB(&callerASXB, NULL);
-
-    callerACEEAddr = callerASXB.asxbsenv;
-
-    copyWithSourceKey(content, callerACEEAddr, 0, CROSS_MEMORY_ALET_HASN, sizeof(ACEE));
-
-  }
-
-  if (address != NULL) {
-    *address = callerACEEAddr;
-  }
-
-}
-
-static void getCallerAddressSpaceACEE(ACEE *content, ACEE **address) {
-
-  ASXB callerASXB;
-  getCallerASXB(&callerASXB, NULL);
-
-  ACEE *callerACEEAddr = callerASXB.asxbsenv;
-
-  copyWithSourceKey(content, callerACEEAddr, 0, CROSS_MEMORY_ALET_HASN, sizeof(ACEE));
-
-  if (address != NULL) {
-    *address = callerACEEAddr;
-  }
-
-}
-
-void cmGetCallerTaskACEE(ACEE *content, ACEE **address) {
-
-  TCB callerTCB;
-  TCB *callerTCBAddress = NULL;
-  getCallerTCB(&callerTCB, &callerTCBAddress);
-
-  ACEE * __ptr32 callerACEEAddr = NULL;
-  if (callerTCBAddress != NULL) {
-      callerACEEAddr = callerTCB.tcbsenv;
-  }
-
-  if (callerACEEAddr != NULL) {
-    copyWithSourceKey(content, callerACEEAddr, 0, CROSS_MEMORY_ALET_HASN,
-                      sizeof(ACEE));
-  }
-
-  if (address != NULL) {
-    *address = callerACEEAddr;
-  }
-
 }
 
 __asm("GLBRLSTP RACROUTE REQUEST=LIST,MF=L" : "DS"(GLBRLSTP));
@@ -1616,7 +1119,7 @@ static bool isCallerAuthorized(CrossMemoryServerGlobalArea *globalArea) {
 
   ACEE callerACEE;
   ACEE *callerACEEAddr = NULL;
-  getCallerAddressSpaceACEE(&callerACEE, &callerACEEAddr);
+  cmGetCallerAddressSpaceACEE(&callerACEE, &callerACEEAddr);
   if (callerACEEAddr == NULL) {
     return FALSE;
   }
@@ -2121,90 +1624,24 @@ static int handlePCCP() {
 }
 
 static void *allocateECSAStorage(unsigned int size) {
-  char *data = NULL;
-  __asm(
-      ASM_PREFIX
-      "         STORAGE OBTAIN"
-      ",COND=YES"
-      ",LENGTH=(%1)"
-      ",SP=228"
-      ",LOC=31"
-      ",OWNER=SYSTEM"
-      ",CALLRKY=YES"
-      ",LINKAGE=SYSTEM                                                         \n"
-      "         LGR   %0,1                                                     \n"
-      : "=r"(data)
-      : "r"(size)
-      : "r0", "r1", "r14", "r15"
-  );
-  return data;
+  return cmAlloc(size, CROSS_MEMORY_SERVER_SUBPOOL, CROSS_MEMORY_SERVER_KEY);
 }
 
 static void freeECSAStorage(void *data, unsigned int size) {
-  __asm(
-      ASM_PREFIX
-      "         STORAGE RELEASE"
-      ",COND=NO"
-      ",LENGTH=(%0)"
-      ",SP=228"
-      ",ADDR=(%1)"
-      ",CALLRKY=YES"
-      ",LINKAGE=SYSTEM                                                         \n"
-      :
-      : "r"(size), "r"(data)
-      : "r0", "r1", "r14", "r15"
-  );
+  cmFree(data, size, CROSS_MEMORY_SERVER_SUBPOOL, CROSS_MEMORY_SERVER_KEY);
 }
 
 static void allocateECSAStorage2(unsigned int size,
                                  void **resultPtr) {
 
-  // Note: Used ADDR= option so STORAGE macro stores the address (for use by
-  // recovery) a few instructions sooner than the compiler would.
-#ifndef _LP64
-#define ROFF 0
-#else
-#define ROFF 4
-#endif
-
-  __asm(
-      ASM_PREFIX
-      "         STORAGE OBTAIN"
-      ",COND=YES"
-      ",LENGTH=(%[size])"
-      ",ADDR="INT_TO_STR(ROFF)"(,%[resultPtr])"
-      ",SP=228"
-      ",LOC=31"
-      ",OWNER=SYSTEM"
-      ",CALLRKY=YES"
-      ",LINKAGE=SYSTEM                                                         \n"
-      : [result]"=m"(*resultPtr)
-      : [size]"NR:r0"(size), [resultPtr]"r"(resultPtr)
-      : "r0", "r1", "r14", "r15"
-  );
+  cmAlloc2(size, CROSS_MEMORY_SERVER_SUBPOOL, CROSS_MEMORY_SERVER_KEY,
+           resultPtr);
 
 }
 
 static void freeECSAStorage2(void **dataPtr, unsigned int size) {
 
-  void *localData = *dataPtr;
-  *dataPtr = NULL;
-
-  __asm(
-      ASM_PREFIX
-      "         STORAGE RELEASE"
-      ",COND=NO"
-      ",LENGTH=(%[size])"
-      ",SP=228"
-      ",ADDR=(%[storageAddr])"
-      ",CALLRKY=YES"
-      ",LINKAGE=SYSTEM                                                         \n"
-      :
-      : [size]"NR:r0"(size), [storageAddr]"r"(localData)
-      : "r0", "r1", "r14", "r15"
-  );
-
-  localData = NULL;
+  cmFree2(dataPtr, size, CROSS_MEMORY_SERVER_SUBPOOL, CROSS_MEMORY_SERVER_KEY);
 
 }
 
@@ -2380,7 +1817,24 @@ static void termStandardServices(CrossMemoryServer *server) {
 
 }
 
-CrossMemoryServer *makeCrossMemoryServer(STCBase *base, const CrossMemoryServerName *name, unsigned int flags, int *reasonCode) {
+CrossMemoryServer *makeCrossMemoryServer(STCBase *base,
+                                         const CrossMemoryServerName *name,
+                                         unsigned int flags, int *reasonCode) {
+
+  return makeCrossMemoryServer2(base, name, flags, NULL, NULL, NULL, reasonCode);
+
+}
+
+CrossMemoryServer *makeCrossMemoryServer2(
+    STCBase *base,
+    const CrossMemoryServerName *name,
+    unsigned int flags,
+    CMSStarCallback *startCallback,
+    CMSStopCallback *stopCallback,
+    void *callbackData,
+    int *reasonCode
+) {
+
 
   int envCheckRC = isEnvironmentReady();
   if (envCheckRC != RC_CMS_OK) {
@@ -2396,6 +1850,10 @@ CrossMemoryServer *makeCrossMemoryServer(STCBase *base, const CrossMemoryServerN
   server->name = name ? *name : CMS_DEFAULT_SERVER_NAME;
   memcpy(server->ddname.text, CMS_DDNAME, sizeof(server->ddname.text));
   memcpy(server->dsname.text, CMS_MODULE, sizeof(server->dsname.text));
+
+  server->startCallback = startCallback;
+  server->stopCallback = stopCallback;
+  server->callbackData = callbackData;
 
   if (flags & CMS_SERVER_FLAG_DEBUG) {
     logSetLevel(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG);
@@ -2469,116 +1927,6 @@ int cmsRegisterService(CrossMemoryServer *server, int id, CrossMemoryServiceFunc
   zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" function %d - 0x%p\n", id, serviceFunction);
 
   return RC_CMS_OK;
-}
-
-__asm("GLBDYLPA    CSVDYLPA MF=(L,GLBDYLPA)" : "DS"(GLBDYLPA));
-
-static int csvdylpaAdd(LPMEA * __ptr32 lpmea, EightCharString  * __ptr32 ddname, EightCharString  * __ptr32 dsname, int *reasonCode) {
-
-  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" about to add %8.8s - %8.8s to LPA\n", ddname->text, dsname->text);
-
-  __asm(" CSVDYLPA MF=L" : "DS"(csvdylpaParmList));
-  csvdylpaParmList = GLBDYLPA;
-
-  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" csvdylpaParmList @ 0x%p, size=%zu\n", &csvdylpaParmList, sizeof(csvdylpaParmList));
-  zowedump(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, (char *)&csvdylpaParmList, sizeof(csvdylpaParmList));
-
-  memset(lpmea, 0, sizeof(LPMEA));
-  memcpy(lpmea->inputInfo.name, dsname->text, 8);
-  lpmea->inputInfo.inputFlags0 |= LPMEA_INPUT_FLAGS0_FIXED;
-
-  char requestor[16];
-  memset(requestor, ' ', sizeof(requestor));
-  memcpy(requestor, "ZSS CM SERVER   ", sizeof(requestor));
-
-  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" lpmea @ 0x%p, size=%zu\n", lpmea, sizeof(LPMEA));
-  zowedump(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, (char *)lpmea, sizeof(LPMEA));
-
-  int moduleCount = 1;
-
-  int rc = 0, rsn = 0;
-  __asm(
-      ASM_PREFIX
-
-#ifdef _LP64
-      "         SAM31                                                          \n"
-      "         SYSSTATE AMODE64=NO                                            \n"
-#endif
-
-      "         CSVDYLPA REQUEST=ADD,MODINFOTYPE=MEMBERLIST,"
-      "MODINFO=(%2),NUMMOD=(%3),"
-      "DDNAME=(%4),"
-      "SECMODCHECK=NO,"
-      "REQUESTOR=(%5),"
-      "RETCODE=%0,RSNCODE=%1,"
-      "MF=(E,(%6))\n"
-
-#ifdef _LP64
-      "         SAM64                                                          \n"
-      "         SYSSTATE AMODE64=YES                                           \n"
-#endif
-
-      : "=m"(rc), "=m"(rsn)
-      : "r"(lpmea), "r"(&moduleCount), "r"(&ddname->text), "r"(&requestor), "r"(&csvdylpaParmList)
-      : "r0", "r1", "r14", "r15"
-  );
-
-  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" lpmea rc=%d, rsn=%d\n", rc, rsn);
-  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" lpmea @ 0x%p, size=%zu\n", lpmea, sizeof(LPMEA));
-  zowedump(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, (char *)lpmea, sizeof(LPMEA));
-
-  *reasonCode = rsn;
-  return rc;
-}
-
-static int csvdylpaDelete(LPMEA * __ptr32 lpmea, int *reasonCode) {
-
-  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" about to delete using LPMEA @ 0x%p\n", lpmea);
-  zowedump(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, (char *)lpmea, sizeof(LPMEA));
-
-  __asm(" CSVDYLPA MF=L" : "DS"(csvdylpaParmList));
-  csvdylpaParmList = GLBDYLPA;
-
-  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" csvdylpaParmList @ 0x%p, size=%zu\n", &csvdylpaParmList, sizeof(csvdylpaParmList));
-  zowedump(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, (char *)&csvdylpaParmList, sizeof(csvdylpaParmList));
-
-  LPMED lpmed;
-  memset(&lpmed, 0, sizeof(LPMED));
-  memcpy(lpmed.inputInfo.name, lpmea->inputInfo.name, sizeof(lpmed.inputInfo.name));
-  memcpy(lpmed.inputInfo.deleteToken, lpmea->outputInfo.stuff.successInfo.deleteToken, sizeof(lpmed.inputInfo.deleteToken));
-
-  int moduleCount = 1;
-
-  int rc = 0, rsn = 0;
-  __asm(
-      ASM_PREFIX
-
-#ifdef _LP64
-      "         SAM31                                                          \n"
-      "         SYSSTATE AMODE64=NO                                            \n"
-#endif
-
-      "         CSVDYLPA REQUEST=DELETE,"
-      "MODINFO=(%2),NUMMOD=(%3),SECMODCHECK=NO,"
-      "RETCODE=%0,RSNCODE=%1,"
-      "MF=(E,(%4))\n"
-
-#ifdef _LP64
-      "         SAM64                                                          \n"
-      "         SYSSTATE AMODE64=YES                                           \n"
-#endif
-
-      : "=m"(rc), "=m"(rsn)
-      : "r"(&lpmed), "r"(&moduleCount), "r"(&csvdylpaParmList)
-      : "r0", "r1", "r14", "r15"
-  );
-
-  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" lpmed rc=%d, rsn=%d\n", rc, rsn);
-  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" lpmed @ 0x%p, size=%zu\n", &lpmed, sizeof(LPMED));
-  zowedump(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, (char *)&lpmed, sizeof(LPMED));
-
-  *reasonCode = rsn;
-  return rc;
 }
 
 static int loadServer(CrossMemoryServer *server) {
@@ -2850,7 +2198,7 @@ static int addGlobalArea(const CrossMemoryServerName *serverName, CrossMemorySer
 
   int unlockRC = 0, unlockRSN = 0;
   unlockRC = isgenqReleaseLock(&lockToken, &unlockRSN);
-  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" RTVE ISGENQ RELEASE RC = %d, RSN = 0x%08X:\n", unlockRC, unlockRSN);
+  zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_DEBUG, CMS_LOG_DEBUG_MSG_ID" ZVTE ISGENQ RELEASE RC = %d, RSN = 0x%08X:\n", unlockRC, unlockRSN);
   if (unlockRC > 4) {
     status = (status != RC_CMS_OK) ? status : RC_CMS_ZVTE_CHAIN_NOT_RELEASED;
   }
@@ -2923,8 +2271,8 @@ static int allocateGlobalResources(CrossMemoryServer *server) {
     memset(globalArea, 0, sizeof(CrossMemoryServerGlobalArea));
     memcpy(globalArea->eyecatcher, CMS_GLOBAL_AREA_EYECATCHER, sizeof(globalArea->eyecatcher));
     globalArea->version = CROSS_MEMORY_SERVER_VERSION;
-    globalArea->key = 4;
-    globalArea->subpool = 228;
+    globalArea->key = CROSS_MEMORY_SERVER_KEY;
+    globalArea->subpool = CROSS_MEMORY_SERVER_SUBPOOL;
     globalArea->size = sizeof(CrossMemoryServerGlobalArea);
 
     int setRC = addGlobalArea(&server->name, globalArea);
@@ -2971,8 +2319,8 @@ static int allocateGlobalResources(CrossMemoryServer *server) {
         zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_INFO, CMS_LOG_DEBUG_MSG_ID
                 " LPA dev mode enabled, issuing CSVDYLPA DELETE\n");
         int lpaDeleteRSN = 0;
-        int lpaDeleteRC = csvdylpaDelete(&globalArea->lpaModuleInfo,
-                                         &lpaDeleteRSN);
+        int lpaDeleteRC = lpaDelete(&globalArea->lpaModuleInfo,
+                                    &lpaDeleteRSN);
         if (lpaDeleteRC != 0) {
           zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_SEVERE,
                   CMS_LOG_LPA_DELETE_FAILURE_MSG, lpaDeleteRC, lpaDeleteRSN);
@@ -3004,7 +2352,8 @@ static int allocateGlobalResources(CrossMemoryServer *server) {
   if (moduleAddressLPA == NULL) {
 
     int lpaAddRSN = 0;
-    int lpaAddRC = csvdylpaAdd(&server->lpaCodeInfo, &server->ddname, &server->dsname, &lpaAddRSN);
+    int lpaAddRC = lpaAdd(&server->lpaCodeInfo, &server->ddname, &server->dsname,
+                          &lpaAddRSN);
     if (lpaAddRC != 0) {
       zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_SEVERE, CMS_LOG_LPA_LOAD_FAILURE_MSG, lpaAddRC, lpaAddRSN);
       return RC_CMS_LPA_ADD_FAILED;
@@ -3951,6 +3300,7 @@ int cmsStartMainLoop(CrossMemoryServer *srv) {
 
   int status = RC_CMS_OK;
   bool serverStarted = false;
+  bool startCallbackSuccess = true;
   bool resourceManagerInstalled = false;
 
   if (status == RC_CMS_OK) {
@@ -4038,6 +3388,18 @@ int cmsStartMainLoop(CrossMemoryServer *srv) {
   }
 
   if (status == RC_CMS_OK) {
+    if (srv->startCallback != NULL) {
+      int callbackRC = srv->startCallback(srv->globalArea, srv->callbackData);
+      if (callbackRC != 0) {
+        zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_SEVERE, CMS_LOG_TMP_DEV_MSG
+               "Start callback failed, RC = %d", callbackRC);
+        status = RC_CMS_ERROR;
+        startCallbackSuccess = false;
+      }
+    }
+  }
+
+  if (status == RC_CMS_OK) {
     int startRC = startServer(srv);
     if (startRC != RC_CMS_OK) {
       zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_SEVERE, CMS_LOG_INIT_STEP_FAILURE_MSG, "start server", startRC);
@@ -4093,6 +3455,17 @@ int cmsStartMainLoop(CrossMemoryServer *srv) {
     }
   }
 
+  if (startCallbackSuccess) {
+    if (srv->stopCallback != NULL) {
+      int callbackRC = srv->stopCallback(srv->globalArea, srv->callbackData);
+      if (callbackRC != 0) {
+        zowelog(NULL, LOG_COMP_ID_CMS, ZOWE_LOG_SEVERE, CMS_LOG_TMP_DEV_MSG
+               "Stop callback failed, RC = %d", callbackRC);
+        status = RC_CMS_ERROR;
+      }
+    }
+  }
+
   if (resourceManagerInstalled) {
     int rmDeleteRC = removeResourceManager(&rmHandle);
     if (rmDeleteRC != RC_RESMGR_OK) {
@@ -4110,7 +3483,7 @@ int cmsStartMainLoop(CrossMemoryServer *srv) {
   return status;
 }
 
-#endif /* METTLE */
+#endif /* defined(METTLE) && !defined(CMS_CLIENT) */
 
 int cmsGetGlobalArea(const CrossMemoryServerName *serverName, CrossMemoryServerGlobalArea **globalAreaAddress) {
 
@@ -4243,6 +3616,20 @@ int cmsCallService(const CrossMemoryServerName *serverName, int serviceID, void 
   return callServiceInternal(globalArea, serviceID, parmList, serviceRC);
 }
 
+int cmsCallService2(CrossMemoryServerGlobalArea *cmsGlobalArea,
+                    int serviceID, void *parmList, int *serviceRC) {
+
+  if (CROSS_MEMORY_SERVER_MAX_SERVICE_ID < serviceID) {
+    return RC_CMS_FUNCTION_ID_OUT_OF_RANGE;
+  }
+
+  if (cmsGlobalArea->version != CROSS_MEMORY_SERVER_VERSION) {
+    return RC_CMS_WRONG_SERVER_VERSION;
+  }
+
+  return callServiceInternal(cmsGlobalArea, serviceID, parmList, serviceRC);
+}
+
 int cmsPrintf(const CrossMemoryServerName *serverName, const char *formatString, ...) {
 
   CrossMemoryServerLogServiceParm msgParmList;
@@ -4318,7 +3705,7 @@ CrossMemoryServerName cmsMakeServerName(const char *nameNullTerm) {
   return name;
 }
 
-#ifdef METTLE
+#if defined(METTLE) && !defined(CMS_CLIENT)
 
 #define crossmemoryServerDESCTs CMSDSECT
 void crossmemoryServerDESCTs(){
@@ -4405,8 +3792,7 @@ void crossmemoryServerDESCTs(){
 
 }
 
-#endif /* METTLE */
-
+#endif /* defined(METTLE) && !defined(CMS_CLIENT) */
 
 /*
   This program and the accompanying materials are
@@ -4417,4 +3803,3 @@ void crossmemoryServerDESCTs(){
   
   Copyright Contributors to the Zowe Project.
 */
-
