+++ h/logging.h
@@ -72,6 +72,11 @@ ZOWE_PRAGMA_PACK_RESET
  * */
 #define LOG_ZOWE_VENDOR_ID     0x008F
 
+#define LOG_PROD_COMMON        0x008D000100000000LLU
+#define LOG_PROD_ZSS           0x008D000200000000LLU
+
+#define LOG_PROD_CMS           0x008D000400000000LLU
+
 #define LOG_COMP_ALLOC         0x008F000100010000LLU
 #define LOG_COMP_UTILS         0x008F000100020000LLU
 #define LOG_COMP_COLLECTIONS   0x008F000100030000LLU
@@ -83,6 +88,7 @@ ZOWE_PRAGMA_PACK_RESET
 #define LOG_COMP_HTTPSERVER    0x008F000100090000LLU
 #define LOG_COMP_DISCOVERY     0x008F0001000A0000LLU
 #define LOG_COMP_DATASERVICE   0x008F0001000B0000LLU
+#define LOG_COMP_LPA           0x008F0001000C0000LLU
 
 #define LOG_DEST_DEV_NULL      0x008F0000
 #define LOG_DEST_PRINTF_STDOUT 0x008F0001
