+++ h/zowetypes.h
@@ -356,6 +356,11 @@ typedef unsigned long long uint64;
 #define PAD_LONG(x,y) int filler##x; y
 #endif
 
+/* TODO better place for such a type? */
+typedef struct EightCharString_tag {
+  char text[8];
+} EightCharString;
+
 #endif
